# src/main/resources/openapi/contracts/meetingService.yaml
openapi: 3.0.0
info:
  title: Meeting Management API
  description: API for managing meetings and participants
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
  /participants:
    get:
      tags:
        - participant-query-controller
      summary: Get all participants
      operationId: getParticipants
      responses:
        '200':
          description: List of participants
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Participant'
    post:
      tags:
        - participant-command-controller
      summary: Create a new participant
      operationId: saveParticipant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Participant'
      responses:
        '201':
          description: Participant created successfully
          content:
            'application/json':
              schema:
                type: object
                $ref: '#/components/schemas/Participant'
        '400':
          description: Invalid input
          content:
            'application/json':
              schema:
                type: object
                $ref: '#/components/schemas/ApiError'
  /participants/{id}:
    delete:
      tags:
        - participant-command-controller
      summary: Delete a participant
      operationId: deleteParticipant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: long
      responses:
        '204':
          description: Participant deleted successfully
...
components:
  schemas:
    Participant:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: long
        name:
          type: string
        email:
          type: string
...
    ApiError:
      type: object
      required:
        - cause
      properties:
        cause:
          type: string